+++ src/main/java/org/truffleruby/core/format/printf/PrintfSimpleTreeBuilder.java
@@ -258,7 +258,7 @@ public final class PrintfSimpleTreeBuilder {
                         break;
                     default:
                         throw new UnsupportedOperationException(
-                                "unsupported type: " + config.getFormatType().toString());
+                                "unsupported type: " + config.getFormatType());
                 }
 
             }