+++ src/main/java/org/truffleruby/core/objectspace/ObjectSpaceNodes.java
@@ -25,6 +25,7 @@ import org.truffleruby.core.numeric.BigIntegerOps;
 import org.truffleruby.core.numeric.RubyBignum;
 import org.truffleruby.core.proc.RubyProc;
 import org.truffleruby.core.string.StringUtils;
+import org.truffleruby.core.support.TypeNodes;
 import org.truffleruby.core.symbol.RubySymbol;
 import org.truffleruby.language.ImmutableRubyObject;
 import org.truffleruby.language.NotProvided;
@@ -279,9 +280,12 @@ public abstract class ObjectSpaceNodes {
 
         @TruffleBoundary
         @Specialization
-        Object undefineFinalizer(RubyDynamicObject object) {
+        Object undefineFinalizer(RubyDynamicObject object,
+                @Cached TypeNodes.CheckFrozenNode raiseIfFrozenNode) {
             final DynamicObjectLibrary objectLibrary = DynamicObjectLibrary.getUncached();
             synchronized (object) {
+                raiseIfFrozenNode.execute(this, object);
+
                 FinalizerReference ref = (FinalizerReference) objectLibrary
                         .getOrDefault(object, Layouts.FINALIZER_REF_IDENTIFIER, null);
                 if (ref != null) {