+++ src/main/java/org/truffleruby/language/constants/LookupConstantBaseNode.java
@@ -10,45 +10,27 @@
 package org.truffleruby.language.constants;
 
 import com.oracle.truffle.api.nodes.Node;
-import org.truffleruby.core.module.ModuleOperations;
 import org.truffleruby.core.module.RubyModule;
 import org.truffleruby.language.RubyBaseNode;
 import org.truffleruby.language.WarnNode;
 
 import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
-import com.oracle.truffle.api.source.SourceSection;
 
 public abstract class LookupConstantBaseNode extends RubyBaseNode {
 
-    @Child private WarnNode warnNode;
+    @Child private WarnDeprecatedConstantNode warnDeprecatedConstantNode;
 
     protected void warnDeprecatedConstant(RubyModule module, String name) {
-        if (warnNode == null) {
+        if (warnDeprecatedConstantNode == null) {
             CompilerDirectives.transferToInterpreterAndInvalidate();
-            warnNode = insert(new WarnNode());
+            warnDeprecatedConstantNode = insert(new WarnDeprecatedConstantNode());
         }
 
-        if (warnNode.shouldWarnForDeprecation()) {
-            warnNode.warningMessage(getSection(this), formatMessage(this, module, name));
-        }
+        warnDeprecatedConstantNode.warnDeprecatedConstant(module, name);
     }
 
     protected static void warnDeprecatedConstant(Node node, WarnNode warnNode, RubyModule module, String name) {
-        if (warnNode.shouldWarnForDeprecation()) {
-            warnNode.warningMessage(getSection(node), formatMessage(node, module, name));
-        }
-    }
-
-    @TruffleBoundary
-    private static SourceSection getSection(Node node) {
-        return getContext(node).getCallStack()
-                .getTopMostUserSourceSection(getAdoptedNode(node).getEncapsulatingSourceSection());
-    }
-
-    @TruffleBoundary
-    private static String formatMessage(Node node, RubyModule module, String name) {
-        return "constant " + ModuleOperations.constantName(getContext(node), module, name) + " is deprecated";
+        WarnDeprecatedConstantNode.warnDeprecatedConstant(node, warnNode, module, name);
     }
 
     protected int getCacheLimit() {